-- convert API fixed

+++ Shared/Models/CurrenciesConvertDetails.cs
@@ -2,13 +2,9 @@
 {
     public class CurrenciesConvertDetails
     {
-        /// <summary>
-        /// Արժույթը՝ դեպի որը կատարվում է փոխարկումը
-        /// </summary>
+       
         public string To { get; set; }
-        /// <summary>
-        /// Գումարի չափը, որը ստացվում է փոխարկման արդյունքում 
-        /// </summary>
-        public double Value { get; set; }
+
+        public decimal Value { get; set; }
     }
 }
